+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 next:
 * Attribute mdl-layout-fixed-drawer for mdl-layout
 * Attribute mdl-layout-fixed-header for mdl-layout
+* Mode scroll, waterfall for mdl-layout
 
 <a name="1.0.0-alpha.1"></a>
 # 1.0.0-alpha.1 (2016-07-06)